REDIS 测试

+++ src/main/java/com/ycmm/common/cache/impl/EhcacheImpl.java
@@ -136,7 +136,7 @@ public class EhcacheImpl implements CacheService {
 	}
 
     @Override
-    public Integer incr(String key) {
+    public Long incr(String key) {
         return null;
     }
